+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+2013-09-25
+----------
+
+  * Added view helper to have an ability to include asset collections in the view script manually.
+  
+    Only the collection name is required.
+    
+    ```
+    <!--[if IE 7]>
+        <?php echo $this->asset('login_ie7_head_js', array('type' => 'text/javascript')) ?>
+        <?php echo $this->asset('login_ie7_css', array('media' => 'screen', 'type' => 'screen', 'text/css' => 'stylesheet')) ?>
+    <![endif]-->
+    ```
+    
+    Register view helper in your module if necessary:
+    
+    ```php
+    <?php
+    public function getViewHelperConfig()
+    {
+        return array(
+            'factories' => array(
+                'asset' => function($helpers) {
+                    /** @var HelperPluginManager $helpers */
+
+                    $sl = $helpers->getServiceLocator();
+                    return new Helper\Asset($sl);
+                },
+            ),
+        );
+    }
+    ```
+
 2013-07-25
 ----------